Only report on 2 sparks per period.
This is to limit the amount of time taken away from robot code. The actual number of sparks reported on can be adjusted easily later if necessary. 2 was chosen because even if there are 20 sparks (ie every PDH slot is used), they would each still be checked every 0.2 secs at the default period of 20ms. A NEO 550 would take about 2 seconds to burnout, so this provides a 10x factor of safety.